Extract FieldNoteExport.ExportTask
