+++ main/src/main/java/cgeo/geocaching/export/FieldNoteExport.java
@@ -1,29 +1,16 @@
 package cgeo.geocaching.export;
 
-import cgeo.geocaching.CgeoApplication;
 import cgeo.geocaching.R;
-import cgeo.geocaching.activity.ActivityMixin;
-import cgeo.geocaching.connector.ConnectorFactory;
-import cgeo.geocaching.connector.IConnector;
-import cgeo.geocaching.connector.capability.FieldNotesCapability;
-import cgeo.geocaching.log.LogEntry;
 import cgeo.geocaching.models.Geocache;
 import cgeo.geocaching.settings.Settings;
-import cgeo.geocaching.storage.ContentStorage;
-import cgeo.geocaching.storage.DataStore;
 import cgeo.geocaching.storage.PersistableFolder;
 import cgeo.geocaching.ui.dialog.Dialogs;
-import cgeo.geocaching.utils.AsyncTaskWithProgress;
 import cgeo.geocaching.utils.Formatter;
-import cgeo.geocaching.utils.Log;
-import cgeo.geocaching.utils.ShareUtils;
 import cgeo.geocaching.utils.UriUtils;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.Dialog;
-import android.content.Context;
-import android.net.Uri;
 import android.view.View;
 import android.widget.CheckBox;
 import android.widget.TextView;
@@ -32,7 +19,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 
-import java.io.File;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
@@ -42,7 +28,6 @@ import java.util.Locale;
  * Exports offline logs in the Groundspeak Field Note format.
  */
 public class FieldNoteExport extends AbstractExport {
-    private static int fieldNotesCount = 0;
     private final String fileName;
 
     public FieldNoteExport() {
@@ -53,11 +38,11 @@ public class FieldNoteExport extends AbstractExport {
 
     @Override
     public void export(@NonNull final List<Geocache> cachesList, @Nullable final Activity activity) {
-        final Geocache[] caches = cachesList.toArray(new Geocache[cachesList.size()]);
+        final Geocache[] caches = cachesList.toArray(new Geocache[0]);
         if (activity == null) {
             // No activity given, so no user interaction possible.
             // Start export with default parameters.
-            new ExportTask(null, false, false).execute(caches);
+            new FieldNoteExportExportTask(null, false, false, getProgressTitle(), fileName, getName()).execute(caches);
         } else {
             // Show configuration dialog
             getExportOptionsDialog(caches, activity).show();
@@ -90,113 +75,10 @@ public class FieldNoteExport extends AbstractExport {
             Settings.setFieldNoteExportUpload(upload);
             Settings.setFieldNoteExportOnlyNew(onlyNew);
             dialog.dismiss();
-            new ExportTask(activity, upload, onlyNew).execute(caches);
+            new FieldNoteExportExportTask(activity, upload, onlyNew, getProgressTitle(), fileName, getName()).execute(caches);
         });
 
         return builder.create();
-    }
-
-    private class ExportTask extends AsyncTaskWithProgress<Geocache, Boolean> {
-        private final boolean upload;
-        private final boolean onlyNew;
-        private Uri exportUri;
-
-        private static final int STATUS_UPLOAD = -1;
-
-        /**
-         * Instantiates and configures the task for exporting field notes.
-         *
-         * @param activity optional: Show a progress bar and toasts
-         * @param upload   Upload the Field Note to geocaching.com
-         * @param onlyNew  Upload/export only new logs since last export
-         */
-        ExportTask(@Nullable final Activity activity, final boolean upload, final boolean onlyNew) {
-            super(activity, getProgressTitle(), CgeoApplication.getInstance().getString(R.string.export_fieldnotes_creating), true);
-            this.upload = upload;
-            this.onlyNew = onlyNew;
-        }
-
-        @Override
-        protected Boolean doInBackgroundInternal(final Geocache[] caches) {
-            // export all field notes, without any filtering by connector
-            final FieldNotes fieldNotes = createFieldNotes(caches);
-            if (fieldNotes == null) {
-                return false;
-            }
-
-            // write to uri
-            exportUri = fieldNotes.writeToFolder(PersistableFolder.FIELD_NOTES.getFolder(), fileName);
-            if (exportUri == null) {
-                return false;
-            }
-            fieldNotesCount = fieldNotes.size();
-            // upload same file to multiple connectors, if they support the upload
-            return uploadFieldNotes();
-        }
-
-        private Boolean uploadFieldNotes() {
-            boolean uploadResult = true;
-            if (upload) {
-                publishProgress(STATUS_UPLOAD);
-                final File tempFile = ContentStorage.get().writeUriToTempFile(exportUri, fileName);
-                if (tempFile != null) {
-                    for (final IConnector connector : ConnectorFactory.getConnectors()) {
-                        if (connector instanceof FieldNotesCapability) {
-                            uploadResult &= ((FieldNotesCapability) connector).uploadFieldNotes(tempFile);
-                        }
-                    }
-                    if (!tempFile.delete()) {
-                        Log.i("Temp file could not be deleted: " + tempFile);
-                    }
-                }
-            }
-            return uploadResult;
-        }
-
-        @Nullable
-        private FieldNotes createFieldNotes(final Geocache[] caches) {
-            final FieldNotes fieldNotes = new FieldNotes();
-            try {
-                for (final Geocache cache : caches) {
-                    if (cache.hasLogOffline()) {
-                        final LogEntry log = DataStore.loadLogOffline(cache.getGeocode());
-                        if (log != null && (!onlyNew || log.date > Settings.getFieldnoteExportDate())) {
-                            fieldNotes.add(cache, log);
-                        }
-                    }
-                    publishProgress(fieldNotes.size());
-                }
-            } catch (final Exception e) {
-                Log.e("FieldNoteExport.ExportTask generation", e);
-                return null;
-            }
-            return fieldNotes;
-        }
-
-        @Override
-        protected void onPostExecuteInternal(final Boolean result) {
-            if (activity != null) {
-                final Context nonNullActivity = activity;
-                if (result && exportUri != null) {
-                    Settings.setFieldnoteExportDate(System.currentTimeMillis());
-
-                    ShareUtils.shareOrDismissDialog(activity, exportUri, "text/plain", R.string.export, getName() + " " + nonNullActivity.getString(R.string.export_exportedto) + ": " + UriUtils.toUserDisplayableString(exportUri));
-
-                    if (upload) {
-                        ActivityMixin.showToast(activity, nonNullActivity.getString(R.string.export_fieldnotes_upload_success));
-                    }
-                } else {
-                    ActivityMixin.showToast(activity, nonNullActivity.getString(R.string.export_failed));
-                }
-            }
-        }
-
-        @Override
-        protected void onProgressUpdateInternal(final Integer status) {
-            if (activity != null) {
-                setMessage(activity.getString(status == STATUS_UPLOAD ? R.string.export_fieldnotes_uploading : R.string.export_fieldnotes_creating) + " (" + fieldNotesCount + ')');
-            }
-        }
     }
 
 }
